+++ frontend/index.html
@@ -6,12 +6,11 @@
     <title>Do you remember the way out</title>
     <link rel="stylesheet" href="./styles/styles.css" />
     <link rel="stylesheet" href="./styles/maze-selection.css" />
-    <link rel="stylesheet" href="./styles/maze-details.css" />
     <link rel="stylesheet" href="./styles/main-page.css" />
     <link rel="stylesheet" href="./styles/login.css" />
     <link rel="stylesheet" href="./styles/error.css" />
     <link rel="stylesheet" href="./styles/username.css" />
-    <link rel="stylesheet" href="./styles/maze-game-temp.css" />
+    <link rel="stylesheet" href="./styles/maze-game.css" />
     <link rel="stylesheet" href="./styles/leaderboard.css" />
     <script defer type="module" src="./app.js"></script>
   </head>
