+++ frontend/scripts/maze-game.js
