+++ frontend/styles/leaderboard.css
@@ -120,3 +120,226 @@ td {
   box-shadow: 0 0.04rem 0.13rem #0001;
   transform: scale(0.97);
 }
+
+.maze-details {
+  border: 3px solid var(--accent);
+  background: var(--primary);
+  border-radius: 1.5em;
+  box-shadow: 0 4px 24px rgba(0, 0, 0, 0.1);
+  padding: 2.5em 2em 2em 2em;
+  min-width: 320px;
+  max-width: 480px;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 1.5em;
+}
+
+header h1 {
+  text-align: center;
+  font-size: 2rem;
+  color: #fff;
+  margin: 0 0 0.5em 0;
+  font-weight: 600;
+  letter-spacing: 0.02em;
+}
+
+aside ul {
+  list-style: disc inside;
+  margin: 0 0 0.5em 0;
+  padding: 0;
+  color: #fff;
+  font-size: 1.1em;
+}
+
+.leaderboard {
+  text-align: center;
+  margin: 1em 0;
+}
+
+.leaderboard h2 {
+  color: #fff;
+  font-size: 1.3em;
+  font-weight: 500;
+  margin: 0;
+  letter-spacing: 0.01em;
+}
+
+footer {
+  display: flex;
+  justify-content: space-between;
+  gap: 1em;
+}
+
+@media (max-width: 500px) {
+  .maze-details {
+    padding: 1.2em 0.5em 1em 0.5em;
+    min-width: unset;
+    max-width: 98vw;
+  }
+  footer {
+    flex-direction: column;
+    gap: 0.7em;
+  }
+}
+
+.leaderboard-table {
+  width: 100%;
+  margin: 1em auto 0 auto;
+  border-collapse: separate;
+  border-spacing: 0;
+  background: #fff;
+  border-radius: 1em;
+  box-shadow: 0 2px 12px rgba(0, 0, 0, 0.07);
+  overflow: hidden;
+  font-size: 1em;
+}
+.leaderboard-table th,
+.leaderboard-table td {
+  padding: 0.85em 1.2em;
+  text-align: left;
+}
+.leaderboard-table th {
+  color: #fff;
+  font-weight: 600;
+  font-size: 1.05em;
+  border-bottom: 2px solid #e0e0e0;
+}
+.leaderboard-table tr {
+  transition: background 0.2s;
+}
+.leaderboard-table tbody tr:nth-child(1) {
+  background: #e3f6ff;
+  font-weight: 600;
+}
+.leaderboard-table tbody tr:nth-child(even) {
+  background: #f7fafc;
+}
+.leaderboard-table tbody tr:hover {
+  background: #e0f2fe;
+}
+.leaderboard-table td {
+  color: var(--text);
+  border-bottom: 1px solid #f0f0f0;
+}
+.leaderboard-table tr:last-child td {
+  border-bottom: none;
+}
+@media (max-width: 500px) {
+  .leaderboard-table th,
+  .leaderboard-table td {
+    padding: 0.6em 0.5em;
+    font-size: 0.98em;
+  }
+}
+
+.maze-details {
+  border: 3px solid var(--accent);
+  background: var(--primary);
+  border-radius: 1.5em;
+  box-shadow: 0 4px 24px rgba(0, 0, 0, 0.1);
+  padding: 2.5em 2em 2em 2em;
+  min-width: 320px;
+  max-width: 480px;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 1.5em;
+}
+
+header h1 {
+  text-align: center;
+  font-size: 2rem;
+  color: #fff;
+  margin: 0 0 0.5em 0;
+  font-weight: 600;
+  letter-spacing: 0.02em;
+}
+
+aside ul {
+  list-style: disc inside;
+  margin: 0 0 0.5em 0;
+  padding: 0;
+  color: #fff;
+  font-size: 1.1em;
+}
+
+.leaderboard {
+  text-align: center;
+  margin: 1em 0;
+}
+
+.leaderboard h2 {
+  color: #fff;
+  font-size: 1.3em;
+  font-weight: 500;
+  margin: 0;
+  letter-spacing: 0.01em;
+}
+
+footer {
+  display: flex;
+  justify-content: space-between;
+  gap: 1em;
+}
+
+@media (max-width: 500px) {
+  .maze-details {
+    padding: 1.2em 0.5em 1em 0.5em;
+    min-width: unset;
+    max-width: 98vw;
+  }
+  footer {
+    flex-direction: column;
+    gap: 0.7em;
+  }
+}
+
+.leaderboard-table {
+  width: 100%;
+  margin: 1em auto 0 auto;
+  border-collapse: separate;
+  border-spacing: 0;
+  border-radius: 1em;
+  box-shadow: 0 2px 12px rgba(0, 0, 0, 0.07);
+  overflow: hidden;
+  font-size: 1em;
+}
+.leaderboard-table th,
+.leaderboard-table td {
+  padding: 0.85em 1.2em;
+  text-align: left;
+}
+.leaderboard-table th {
+  color: #fff;
+  font-weight: 600;
+  font-size: 1.05em;
+  border-bottom: 2px solid #e0e0e0;
+}
+.leaderboard-table tr {
+  transition: background 0.2s;
+}
+.leaderboard-table tbody tr:nth-child(1) {
+  background: #e3f6ff;
+  font-weight: 600;
+}
+.leaderboard-table tbody tr:nth-child(even) {
+  background: #f7fafc;
+}
+.leaderboard-table tbody tr:hover {
+  background: #e0f2fe;
+}
+.leaderboard-table td {
+  color: var(--text);
+  border-bottom: 1px solid #f0f0f0;
+}
+.leaderboard-table tr:last-child td {
+  border-bottom: none;
+}
+@media (max-width: 500px) {
+  .leaderboard-table th,
+  .leaderboard-table td {
+    padding: 0.6em 0.5em;
+    font-size: 0.98em;
+  }
+}
