--- frontend/views/maze-details.html
@@ -1,48 +1,0 @@
-  <section class="maze-details" id="maze-details">
-    <header>
-      <h1>Maze 1</h1>
-    </header>
-    <aside>
-      <ul>
-        <li><strong>Difficulty:</strong> Easy</li>
-        <li><strong>Size:</strong> 5×5</li>
-      </ul>
-    </aside>
-    <section class="leaderboard">
-      <h2>Leaderboard</h2>
-      <table class="leaderboard-table">
-        <thead>
-          <tr>
-            <th>Username</th>
-            <th>Points</th>
-          </tr>
-        </thead>
-        <tbody>
-          <tr>
-            <td>MazeMaster</td>
-            <td>1500</td>
-          </tr>
-          <tr>
-            <td>QuickSolver</td>
-            <td>1200</td>
-          </tr>
-          <tr>
-            <td>PathFinder</td>
-            <td>950</td>
-          </tr>
-          <tr>
-            <td>Speedy</td>
-            <td>800</td>
-          </tr>
-          <tr>
-            <td>Newbie</td>
-            <td>500</td>
-          </tr>
-        </tbody>
-      </table>
-    </section>
-    <footer>
-      <button type="button" id="backBtn">Back to maze selection</button>
-      <button type="button" id="startBtn">Start game</button>
-    </footer>
-  </section>
