--- frontend/views/maze-game-temp.html
@@ -1,26 +1,0 @@
-<section id="game-page">
-  <header id="game-banner">
-    <div class="timer-row">
-      <div class="timer-left">
-        <label for="preview-bar">Preview:</label>
-        <progress id="preview-bar" value="0" max="5"></progress>
-      </div>
-      <div class="timer-center">
-        <p id="steps">Steps:</p>
-      </div>
-      <div class="timer-right">
-        <label for="escape-bar">Escape:</label>
-        <progress id="escape-bar" value="0" max="80"></progress>
-      </div>
-    </div>
-  </header>
-  <section id="grid-container"></section>
-  <footer id="game-footer">
-    <button id="up" class="movement-button">⬆</button>
-    <section>
-      <button id="left" class="movement-button">⬅</button>
-      <button id="down" class="movement-button">⬇</button>
-      <button id="right" class="movement-button">➡</button>
-    </section>
-  </footer>
-</section>
